Add analysis of shooting victims by community area
Add the basics of analyzing shootings victims from NewsroomDB by
community area.

Use PANDAS because I want to learn more about it, and because
it helps when slicing and dicing data.

Use RTree for a spatial index to speed up community area lookups.

Add a number of dependencies for dealing with the spatial data.